# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Host {
  name: string;
  description: string;
}

export interface EpisodeHost {
  name: string;
  description: string;
  /** @nullable */
  voiceId: string | null;
}

export type GenerateEpisodeBodyFormat =
  (typeof GenerateEpisodeBodyFormat)[keyof typeof GenerateEpisodeBodyFormat];

export const GenerateEpisodeBodyFormat = {
  comedy: "comedy",
  debate: "debate",
  explainer: "explainer",
  true_crime: "true_crime",
  hot_takes: "hot_takes",
  interview: "interview",
} as const;

export interface GenerateEpisodeBody {
  /**
   * @minLength 3
   * @maxLength 500
   */
  topic: string;
  format: GenerateEpisodeBodyFormat;
  /**
   * @minItems 2
   * @maxItems 3
   */
  hosts: Host[];
}

export type EpisodeStatus = (typeof EpisodeStatus)[keyof typeof EpisodeStatus];

export const EpisodeStatus = {
  pending: "pending",
  generating: "generating",
  ready: "ready",
  failed: "failed",
} as const;

export interface Episode {
  id: number;
  topic: string;
  format: string;
  hosts: EpisodeHost[];
  /** @nullable */
  audioUrl: string | null;
  /** @nullable */
  scriptText: string | null;
  /** @nullable */
  duration: number | null;
  status: EpisodeStatus;
  /** @nullable */
  errorMessage: string | null;
  createdAt: string;
}

export interface PreviewVoiceBody {
  /**
   * @minLength 20
   * @maxLength 1000
   */
  description: string;
  /** @nullable */
  sampleText?: string | null;
}

export interface VoicePreviewItem {
  generatedVoiceId: string;
  audioBase64: string;
  mediaType: string;
}

export interface VoicePreviewResponse {
  previews: VoicePreviewItem[];
}

export interface SaveVoiceBody {
  generatedVoiceId: string;
  name: string;
  description: string;
}

export interface SavedVoice {
  voiceId: string;
  name: string;
}

export interface HostPreset {
  id: string;
  name: string;
  description: string;
  icon: string;
  suggestedFormats: string[];
  hosts: Host[];
}

export interface FetchUrlBody {
  url: string;
}

export interface FetchUrlResponse {
  topic: string;
  sourceTitle: string;
  sourceUrl: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type GenerateEpisode200 = { [key: string]: unknown };

export type BeginBrowserLoginParams = {
  /**
   * Relative path to redirect to after login (must start with `/`). Defaults to `/`.
   */
  returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
  code?: string;
  state?: string;
  iss?: string;
};