# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: castforge
    description: CastForge podcast generation
  - name: Auth
    description: Browser and mobile authentication endpoints.
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /castforge/episodes:
    get:
      operationId: listEpisodes
      tags: [castforge]
      summary: List all generated episodes
      responses:
        "200":
          description: List of episodes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Episode"

  /castforge/episodes/{id}:
    get:
      operationId: getEpisode
      tags: [castforge]
      summary: Get a specific episode
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Episode details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Episode"
        "404":
          description: Episode not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteEpisode
      tags: [castforge]
      summary: Delete an episode
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted successfully
        "404":
          description: Episode not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /castforge/generate:
    post:
      operationId: generateEpisode
      tags: [castforge]
      summary: Generate a podcast episode (SSE streaming progress)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateEpisodeBody"
      responses:
        "200":
          description: SSE stream of generation progress events
          content:
            text/event-stream:
              schema:
                type: object

  /castforge/voices/preview:
    post:
      operationId: previewVoice
      tags: [castforge]
      summary: Generate voice previews from a natural language description
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PreviewVoiceBody"
      responses:
        "200":
          description: Generated voice previews
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoicePreviewResponse"

  /castforge/voices/save:
    post:
      operationId: saveVoice
      tags: [castforge]
      summary: Save a voice preview to the voice library
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveVoiceBody"
      responses:
        "200":
          description: Saved voice details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedVoice"

  /castforge/fetch-url:
    post:
      operationId: fetchUrlTopic
      tags: [castforge]
      summary: Fetch a URL and extract a podcast topic from its content
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FetchUrlBody"
      responses:
        "200":
          description: Extracted topic and source info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FetchUrlResponse"
        "400":
          description: Invalid URL or fetch failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /castforge/presets:
    get:
      operationId: getHostPresets
      tags: [castforge]
      summary: Get preset host configuration pairs
      responses:
        "200":
          description: Available host presets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HostPreset"

  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'
  /login:
    get:
      tags: [Auth]
      operationId: beginBrowserLogin
      summary: Start the browser OIDC login flow
      parameters:
        - name: returnTo
          in: query
          required: false
          description: Relative path to redirect to after login (must start with `/`). Defaults to `/`.
          schema:
            type: string
      responses:
        '302':
          description: Redirect to the OIDC authorization endpoint.
  /callback:
    get:
      tags: [Auth]
      operationId: handleBrowserLoginCallback
      summary: Complete the browser OIDC login flow
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: iss
          in: query
          required: false
          schema:
            type: string
            format: uri
      responses:
        '302':
          description: Redirect to the `returnTo` path on success (defaults to `/`) or `/api/login` on failure.
  /logout:
    get:
      tags: [Auth]
      operationId: logoutBrowserSession
      summary: Clear the session and begin OIDC logout
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '302':
          description: Redirect to the OIDC provider logout URL.

components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — `Bearer <sid>`.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    Host:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
      required:
        - name
        - description

    EpisodeHost:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        voiceId:
          type: ["string", "null"]
      required:
        - name
        - description
        - voiceId

    GenerateEpisodeBody:
      type: object
      properties:
        topic:
          type: string
          minLength: 3
          maxLength: 500
        format:
          type: string
          enum: [comedy, debate, explainer, true_crime, hot_takes, interview]
        hosts:
          type: array
          items:
            $ref: "#/components/schemas/Host"
          minItems: 2
          maxItems: 3
      required:
        - topic
        - format
        - hosts

    Episode:
      type: object
      properties:
        id:
          type: integer
        topic:
          type: string
        format:
          type: string
        hosts:
          type: array
          items:
            $ref: "#/components/schemas/EpisodeHost"
        audioUrl:
          type: ["string", "null"]
        scriptText:
          type: ["string", "null"]
        duration:
          type: ["integer", "null"]
        status:
          type: string
          enum: [pending, generating, ready, failed]
        errorMessage:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - topic
        - format
        - hosts
        - audioUrl
        - scriptText
        - duration
        - status
        - errorMessage
        - createdAt

    PreviewVoiceBody:
      type: object
      properties:
        description:
          type: string
          minLength: 20
          maxLength: 1000
        sampleText:
          type: ["string", "null"]
      required:
        - description

    VoicePreviewItem:
      type: object
      properties:
        generatedVoiceId:
          type: string
        audioBase64:
          type: string
        mediaType:
          type: string
      required:
        - generatedVoiceId
        - audioBase64
        - mediaType

    VoicePreviewResponse:
      type: object
      properties:
        previews:
          type: array
          items:
            $ref: "#/components/schemas/VoicePreviewItem"
      required:
        - previews

    SaveVoiceBody:
      type: object
      properties:
        generatedVoiceId:
          type: string
        name:
          type: string
        description:
          type: string
      required:
        - generatedVoiceId
        - name
        - description

    SavedVoice:
      type: object
      properties:
        voiceId:
          type: string
        name:
          type: string
      required:
        - voiceId
        - name

    HostPreset:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        icon:
          type: string
        suggestedFormats:
          type: array
          items:
            type: string
        hosts:
          type: array
          items:
            $ref: "#/components/schemas/Host"
      required:
        - id
        - name
        - description
        - icon
        - suggestedFormats
        - hosts

    FetchUrlBody:
      type: object
      properties:
        url:
          type: string
      required:
        - url

    FetchUrlResponse:
      type: object
      properties:
        topic:
          type: string
        sourceTitle:
          type: string
        sourceUrl:
          type: string
      required:
        - topic
        - sourceTitle
        - sourceUrl

    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl]
      properties:
        id:
          type: string
        email:
          type: [string, 'null']
          format: email
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'